// client-app/src/app/layout/models/activity.ts
export interface Activity {
    id: string
    title: string
    date: string
    description: string
    category: string
    city: string
    venue: string
  }

 /*  export type Activity =  {
    id: string
    title: string
    date: string
    description: string
    category: string
    city: string
    venue: string
  } */